move VideoGridViewController back to VideoGridView.swift
